Do not enforce children type when dealing with arrays
This will be properly handled later, for now it should be enough
for most cases and avoid breaking when compiling without `-debug`
(when generating react literals).

When children prop is defined and typed as `Array<Something>`,
each item should be typed as `Something` (if only one item, typing
may need to be tested for both `Array<Something>` and `Something`).

When children prop is not defined, each item should be typed as
`ReactFragment`.